requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
dj-database-url==3.0.0
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
geographiclib==2.0
geopy==2.4.1
googlemap==0.1.0
googlemaps==4.10.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.1
numpy==2.2.6
packaging==25.0
pandas==2.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
razorpay==1.4.2
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
zope.interface==7.2
